Conversion from 'size_t' to 'int' possible loss of data
MessageList size uses internal typedef size_type.

Was mixing unsigned int and size_t, which are different sizes.

size_t should be reserved for storage sizes measured in bytes.

+++ src/flamegpu/gpu/CUDAAgentModel.cu
@@ -156,7 +156,7 @@ void CUDAAgentModel::step(const Simulation& simulation)
 
 	//TODO: simulation.getMaxFunctionsPerLayer()
 	for (unsigned int i = 0; i < simulation.getLayerCount(); i++) {
-		int temp = simulation.getFunctionsAtLayer(i).size();
+		int temp = static_cast<int>(simulation.getFunctionsAtLayer(i).size());
 		nStreams = std::max(nStreams, temp);
 	}
